Fix for https://github.com/dmachard/python-dnsdist-console/issues/4
Parsing the server data based on the header position can lead to truncating values, at least for the Queries counter which is truncated once it reaches 10000000.

The output is space separated, but contains values that can also contain spaces (currently, the server name and pool), making splitting unreliable.
The pool name is output last and spaces in that value don't alter the split.
The server name is output second. To find out if the server name contains spaces, it is retrieved using getServer(); however the value output by showServer() is truncated when longer than 20 chars.

This method of parsing will stop working if additional values containing spaces are added to the output, or if the server id or pool move to a different position.